pyproject.toml
[tool.poetry]
name = "expresso"
version = "0.1.0"
description = ""
authors = ["Qi Zhang <singularitti@outlook.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
attrs = "^19.1"
crystals = "^0.6.4"

[tool.poetry.dev-dependencies]
attrs = "^19.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
